apps/python-server/requirements.txt
sounddevice
numpy
soundfile
pydub
tensorflow
pandas
librosa
SpeechRecognition
resampy==0.4.3
pyAudioAnalysis
hmmlearn
eyed3
imblearn
plotly
mediapipe==0.10.14
opencv-python==4.10.0.84
torch
torchvision
timm==1.0.3
pillow